refactored per Wei's comments
